requirements.txt
Django==2.2.9
djangorestframework
redis
django-redis
gunicorn
psycopg2
celery
celery[redis]
django-celery-beat
django-celery-results
elasticsearch-dsl
django-elasticsearch-dsl
django-elasticsearch-dsl-drf